src/intelligence/risk_engine.py
# Risk Intelligence Engine
from sklearn.ensemble import RandomForestClassifier

def train_risk_model(X, y):
    """Train a risk classification model."""
    model = RandomForestClassifier()
    model.fit(X, y)
    return model