requirements.txt
# Data processing
pandas
numpy
pyspark

# Machine Learning
scikit-learn
xgboost

# Forecasting
statsmodels
prophet

# Visualization
matplotlib
seaborn
streamlit
folium

# API requests
requests

# Database connectors
sqlalchemy
psycopg2

# Workflow management
apache-airflow

# Deployment
docker

# Testing and code quality
pytest
black

# Packaging
setuptools

# Editable install (optional)
-e .
